requirements.txt
accelerate==1.7.0
annotated-types==0.7.0
anyio==4.9.0
bidict==0.23.1
bitsandbytes==0.46.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
ConfigArgParse==1.7
contourpy==1.3.2
cycler==0.12.1
durationpy==0.10
fastapi==0.115.13
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
fonttools==4.57.0
fsspec==2025.5.1
gevent==24.11.1
geventhttpclient==2.3.3
google-auth==2.40.3
greenlet==3.2.1
h11==0.16.0
hf-xet==1.1.2
huggingface-hub==0.32.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
kubernetes==33.1.0
llvmlite==0.44.0
locust==2.37.0
locust-cloud==1.21.2
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pydantic==2.11.7
pydantic_core==2.33.2
pylcs==0.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==80.3.1
shap==0.47.2
simple-websocket==1.1.0
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
triton==3.3.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
